Seguidor: separar el componente de vistas para testeo más facil.
