+++ src/Seguidor/Seguidor.jsx
@@ -1,22 +1,7 @@
 import React, { Component } from 'react'
 import FetchingIndicator from '../Fetching/FetchingIndicator'
 import SeguidorHeading from './SeguidorHeading'
-import CarouselView from './CarouselView/CarouselView';
-import TreeView from './TreeView/TreeView';
-import FinalesPendientes from './FinalesPendientes/FinalesPendientes';
-
-const SeguidorView = ({view, materias, updateEstado}) => {
-
-  if(view === 'tree')
-     return <TreeView materias={materias} updateEstado={updateEstado} />;
-
-  if(view === 'carousel')
-     return <CarouselView materias={materias} updateEstado={updateEstado} />;
-
-  if(view === 'finales')
-    return <FinalesPendientes materias={materias} updateEstado={updateEstado} />;
-
-}
+import SeguidorView from './SeguidorView';
 
 class Seguidor extends Component {
